requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2020.12.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.2
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
identify==2.2.5
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
pre-commit==2.13.0
pygraphviz==1.7
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
